+++ src/pages/games/AppleGame/AppleGamePage.tsx
@@ -41,9 +41,24 @@ const AppleGamePage = () => {
       </div>
 
       <div className="mx-auto flex w-full max-w-4xl justify-center gap-8 py-10">
-        <Button onClick={() => setModeState('default')}>일반 모드</Button>
-        <Button onClick={() => setModeState('classic')}>클래식 모드</Button>
-        <Button onClick={() => setModeState('new')}>실험 모드</Button>
+        <Button
+          onClick={() => setModeState('default')}
+          className={modeState === 'default' ? 'bg-green-400' : ''}
+        >
+          일반 모드
+        </Button>
+        <Button
+          onClick={() => setModeState('classic')}
+          className={modeState === 'classic' ? 'bg-green-400' : ''}
+        >
+          클래식 모드
+        </Button>
+        <Button
+          onClick={() => setModeState('new')}
+          className={modeState === 'new' ? 'bg-green-400' : ''}
+        >
+          신규 모드
+        </Button>
       </div>
 
       <Footer />
