+++ src/pages/games/AppleGame/game/viewModal/AppleGameController.tsx
