+++ src/pages/games/AppleGame/game/viewModal/SnackGameController.tsx
